refactor: Move requireModule to utils
